Extract Base64 into _NIODataStructures + formatting
